workflows: moves API to new structure
